hm1

+++ Homework/exercise5_09.09.23.py
@@ -1,8 +1,14 @@
-print ("Введите x"); x=int(input())
-print ("Введите y"); y=int(input())
-if (x > 0 and y > 0) : print("Точка в 1 четверти")
-elif (x < 0 and y>0) : print("Точка в 2 четверти")
-elif (x<0 and y<0) : print("Точка в 3 четверти")
-elif (x>0 and y<0) : print("Точка в 4 четверти")
-elif (x==0 and y!=0) : print("Точка на оси Y")
-elif (x!=0 and y==0) :print("Точка на оси X")
+print ("Введите x"); x=float(input())
+print ("Введите y"); y=float(input())
+if (x > 0 and y > 0) :
+     print("Точка в 1 четверти")
+elif (x < 0 and y>0) :
+     print("Точка в 2 четверти")
+elif (x<0 and y<0) : 
+     print("Точка в 3 четверти")
+elif (x>0 and y<0) :
+     print("Точка в 4 четверти")
+elif (x==0 and y!=0) :
+     print("Точка на оси Y")
+elif (x!=0 and y==0) :
+     print("Точка на оси X")
